Modified StoryAdapter.kt and MainActivity.kt to implement DiffUtil and on click listener for each views.
